pyproject.toml
[project]
name = "expense-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "bcrypt>=5.0.0",
    "flask>=3.1.3",
    "flask-jwt-extended>=4.7.1",
    "flask-restx>=1.3.2",
    "flask-sqlalchemy>=3.1.1",
    "pymysql>=1.1.2",
]
